iota_sdk_transaction_builder::builder::signer

Trait TransactionSigner

pub trait TransactionSigner {
    type Error: 'static + Error + Send + Sync;

    // Required method
    fn sign(
        &self,
        transaction: &Transaction,
    ) -> impl Future<Output = Result<UserSignature, Self::Error>>;
}

Defines a type which can sign a transaction asynchronously.

This trait can be implemented downstream to enable signing when using the TransactionBuilder execute function.

Required Associated Types

type Error: 'static + Error + Send + Sync

The error that can occur during signing.

Required Methods

fn sign( &self, transaction: &Transaction, ) -> impl Future<Output = Result<UserSignature, Self::Error>>

Sign a transaction and return a [UserSignature].

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".