iota_sdk_transaction_builder::builder

Struct TransactionBuilder

#[repr(C)]
pub struct TransactionBuilder<C = (), L = ()> { /* private fields */ }

A transaction builder which can be used to construct [Transaction]s.

Implementations

impl TransactionBuilder

pub fn new(sender: Address) -> Self

Instantiate a new PTB.

pub fn with_client<C>(self, client: C) -> TransactionBuilder<C>

Set the client to enable automatic object resolution.

impl<C, L> TransactionBuilder<C, L>

pub fn apply_argument<P: PTBArgument>(&mut self, param: P) -> Argument

Apply the given parameter and return the generated argument

pub fn apply_arguments<P: PTBArgumentList>(&mut self, param: P) -> Vec<Argument>

Apply the given parameters and return the generated arguments

pub fn get_gas(&self) -> Vec<ObjectId>

Get the current set gas coins.

pub fn gas_budget(&mut self, gas_budget: u64) -> &mut Self

Set the gas budget. Optional.

pub fn gas_price(&mut self, gas_price: u64) -> &mut Self

Set the gas price. Optional.

pub fn sponsor(&mut self, sponsor: Address) -> &mut Self

Set the sponsor. Optional.

pub fn gas_station_sponsor( &mut self, url: Url, ) -> &mut TransactionBuilder<C, GasStationData>

Set the gas station sponsor. Optional.

pub fn expiration(&mut self, expiration: u64) -> &mut Self

Set the expiration. Optional.

pub fn input(&mut self, input: Input) -> Argument

Make a value available to the transaction as an input.

pub fn pure_bytes(&mut self, bytes: Vec<u8>) -> Argument

Add a pure input using the BCS serialized bytes

pub fn pure<T: Serialize>(&mut self, value: T) -> Argument

Add a pure input

pub fn command(&mut self, command: Command) -> Argument

Add a new command to the PTB

pub fn named_command(&mut self, cmd: Command, name: impl NamedResults)

Manually set a command with an optional name

pub fn get_named_result(&self, name: &str) -> Option<Argument>

Get the value for the given string in the named results map

pub fn move_call( &mut self, package_id: impl Into<ObjectId>, module: &str, function: &str, ) -> &mut TransactionBuilder<C, MoveCall>

Begin building a move call.

pub fn transfer_objects<U: PTBArgumentList>( &mut self, recipient: Address, objects: U, ) -> &mut TransactionBuilder<C>

Transfer objects to a recipient address.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, res};
use iota_types::{Address, Digest, ObjectId, ObjectReference, Transaction};

let client = iota_graphql_client::Client::new_devnet();
let sender =
    Address::from_str("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);

builder.transfer_objects(
    Address::from_str("0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900")?,
    (
        // ObjectIds can be passed when a client is provided
        ObjectId::from_str(
            "0xd04077fe3b6fad13b3d4ed0d535b7ca92afcac8f0f2a0e0925fb9f4f0b30c699",
        )?,
        // ObjectReferences are always allowed, though they must be correct
        ObjectReference {
            object_id: ObjectId::from_str(
                "0x8ef4259fa2a3499826fa4b8aebeb1d8e478cf5397d05361c96438940b43d28c9",
            )?,
            digest: Digest::from_str("4jJMQScR4z5kK3vchvDEFYTiCkZPEYdvttpi3iTj1gEW")?,
            version: 435090179,
        },
        // The result of a previous command can also be used
        res("coin"),
    ),
);

let txn: Transaction = builder.finish().await?;

pub fn send_iota<T: PTBArgument>( &mut self, recipient: Address, amount: T, ) -> &mut TransactionBuilder<C>

Send IOTA to a recipient address.

The amount parameter specifies the quantity in NANOS, where 1 IOTA equals 1_000_000_000 NANOS. That amount is split from the gas coin and sent.

Example

use iota_graphql_client::Client;
use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::Address;

let client = Client::new_devnet();
let from_address =
    Address::from_hex("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;
let to_address =
    Address::from_hex("0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900")?;

let mut builder = TransactionBuilder::new(from_address).with_client(client);
builder.send_iota(to_address, 5000000000u64);
let txn = builder.finish().await?;

pub fn send_coins<T: PTBArgumentList, U: PTBArgument>( &mut self, coins: T, recipient: Address, amount: impl Into<Option<U>>, ) -> &mut TransactionBuilder<C>

Transfer some coins to a recipient address. If multiple coins are provided then they will be merged.

The amount parameter specifies the quantity in NANOS, where 1 IOTA equals 1_000_000_000 NANOS. If amount is provided, that amount is split from the provided coins and sent. If amount is None, the entire coins are transferred.

All provided coins must have the same coin type. Mixing coins of different types will result in an error.

If you intend to transfer all provided coins to another address in a single transaction, consider using TransactionBuilder::transfer_objects() instead.

Example

use iota_graphql_client::Client;
use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let client = Client::new_devnet();
let from_address =
    Address::from_hex("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;
let to_address =
    Address::from_hex("0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900")?;

// This is a coin of type
// 0x3358bea865960fea2a1c6844b6fc365f662463dd1821f619838eb2e606a53b6a::cert::CERT
let coin =
    ObjectId::from_hex("0x8ef4259fa2a3499826fa4b8aebeb1d8e478cf5397d05361c96438940b43d28c9")?;

let mut builder = TransactionBuilder::new(from_address).with_client(client);
builder.send_coins([coin], to_address, 50000000000u64);
let txn = builder.finish().await?;

pub fn merge_coins<T: PTBArgument, U: PTBArgumentList>( &mut self, primary_coin: T, consumed_coins: U, ) -> &mut TransactionBuilder<C>

Merge multiple coins into one.

This method combines the balances of multiple coins of the same coin type into a single coin. The primary_coin will receive the balances from all consumed_coins. After merging, the consumed_coins will be consumed and no longer exist.

Example

use iota_graphql_client::Client;
use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let client = Client::new_devnet();
let sender =
    Address::from_hex("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;

let coin_0 =
    ObjectId::from_hex("0x0b0270ee9d27da0db09651e5f7338dfa32c7ee6441ccefa1f6e305735bcfc7ab")?;
let coin_1 =
    ObjectId::from_hex("0xd04077fe3b6fad13b3d4ed0d535b7ca92afcac8f0f2a0e0925fb9f4f0b30c699")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.merge_coins(coin_0, [coin_1]);
let txn = builder.finish().await?;

pub fn split_coins<T: PTBArgument, U: PTBArgumentList>( &mut self, coin: T, split_amounts: U, ) -> &mut TransactionBuilder<C, SplitCoins>

Split a coin into many.

Example

use iota_graphql_client::Client;
use iota_sdk_transaction_builder::{TransactionBuilder, res};
use iota_types::{Address, ObjectId};

let client = Client::new_devnet();
let sender =
    Address::from_hex("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;
let coin =
    ObjectId::from_hex("0x0b0270ee9d27da0db09651e5f7338dfa32c7ee6441ccefa1f6e305735bcfc7ab")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder
    .split_coins(coin, [1000u64, 2000, 3000])
    .name(("coin1", "coin2", "coin3"))
    .transfer_objects(sender, (res("coin1"), res("coin2"), res("coin3")));
let txn = builder.finish().await?;

pub fn publish( &mut self, package_data: MovePackageData, ) -> &mut TransactionBuilder<C, Publish>

Publish a move package.

pub fn upgrade<U: PTBArgument>( &mut self, package_id: ObjectId, package_data: MovePackageData, upgrade_ticket: U, ) -> &mut TransactionBuilder<C, Upgrade>

Upgrade a move package.

pub fn stake<S: PTBArgument>( &mut self, stake_amount: S, validator_address: Address, ) -> &mut Self

Add stake to a validator’s staking pool.

This is a high-level function which will split the provided stake amount from the gas coin and then stake using the resulting coin.

Example

use iota_graphql_client::Client;
use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let client = Client::new_devnet();
let sender =
    Address::from_hex("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;
let validator_address = client
    .active_validators(None, Default::default())
    .await?
    .data
    .into_iter()
    .next()
    .ok_or_else(|| eyre::eyre!("no validators found"))?
    .address
    .address;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.stake(1000000000u64, validator_address);
let txn = builder.finish().await?;

pub fn unstake<S: PTBArgument>(&mut self, staked_iota: S) -> &mut Self

Withdraw stake from a validator’s staking pool.

Example

use iota_graphql_client::Client;
use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let client = Client::new_devnet();
let sender =
    Address::from_hex("0x6f0202b12cd398166bdd3716c9aa3f0b6218ba125491f7ea2bc660fdd5e57ff8")?;
// This is a 0x3::staking_pool::StakedIota owned by the sender
let staked_iota =
    ObjectId::from_hex("0x00030af99878926cd11f8bdf4d2f67c4aa753a4afc249d776c8ed2cc88d7b8d5")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.unstake(staked_iota);
let txn = builder.finish().await?;

pub fn make_move_vec<T: PTBArgument + MoveType>( &mut self, elements: impl IntoIterator<Item = T>, ) -> &mut TransactionBuilder<C, MakeMoveVector>

Make a move vector from a list of elements.

Often it is possible (and more efficient) to pass a rust slice or Vec instead of calling this function, which will serialize the bytes into a move vector pure argument.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, res};
use iota_types::{Address, Transaction};

let client = iota_graphql_client::Client::new_devnet();
let sender = "0x71b4b4f171b4355ff691b7c470579cf1a926f96f724e5f9a30efc4b5f75d085e".parse()?;

let mut builder = TransactionBuilder::new(sender).with_client(client);

let address1 =
    Address::from_str("0xde49ea53fbadee67d3e35a097cdbea210b659676fc680a0b0c5f11d0763d375e")?;
let address2 =
    Address::from_str("0xe512234aa4ef6184c52663f09612b68f040dd0c45de037d96190a071ca5525b3")?;

builder
    .make_move_vec([address1, address2])
    .name("addresses")
    .move_call(Address::FRAMEWORK, "vec_map", "from_keys_values")
    .generics::<(Address, u64)>()
    .arguments((res("addresses"), [10000000u64, 20000000u64]));

let txn: Transaction = builder.finish().await?;

impl<L> TransactionBuilder<(), L>

pub fn gas( &mut self, obj_refs: impl IntoIterator<Item = ObjectReference>, ) -> &mut Self

Add gas coins that will be consumed. Optional.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, res, unresolved};
use iota_types::{Address, Digest, ObjectId, ObjectReference, Transaction};

let sender =
    Address::from_str("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;

let mut builder = TransactionBuilder::new(sender);

let gas_coin1 = ObjectReference {
    object_id: ObjectId::from_str(
        "0x0b0270ee9d27da0db09651e5f7338dfa32c7ee6441ccefa1f6e305735bcfc7ab",
    )?,
    digest: Digest::from_str("CPpQZqyHZcG2Pb9gZyikbc8dEuyipXHR6ihnfe9iYiMt")?,
    version: 473053811,
};
let gas_coin2 = ObjectReference {
    object_id: ObjectId::from_str(
        "0xd04077fe3b6fad13b3d4ed0d535b7ca92afcac8f0f2a0e0925fb9f4f0b30c699",
    )?,
    digest: Digest::from_str("8ahH5RXFnK1jttQEWTypYX7MRzLuQDEXk7fhMHCyZekX")?,
    version: 473053810,
};

builder
    .split_coins(unresolved::Argument::Gas, [1000u64])
    .gas([gas_coin1, gas_coin2])
    .gas_budget(1000000000)
    .gas_price(100);

let txn: Transaction = builder.finish()?;

pub fn finish(self) -> Result<Transaction, Error>

Convert this builder into a transaction.

pub async fn execute_with_gas_station( self, keypair: &SimpleKeypair, ) -> Result<Value, Error>

Execute the transaction using the gas station and return the JSON transaction effects. This will fail unless data is set with Self::gas_station_sponsor.

NOTE: These effects are not necessarily compatible with [TransactionEffects]

impl<L> TransactionBuilder<&Client, L>

pub fn get_client(&self) -> &Client

Get the client used by the builder.

impl<L> TransactionBuilder<Client, L>

pub fn get_client(&self) -> &Client

Get the client used by the builder.

impl<C: ClientMethods, L> TransactionBuilder<C, L>

pub fn gas(&mut self, obj_ids: impl IntoIterator<Item = ObjectId>) -> &mut Self

Add gas coins that will be consumed. If no gas coins are provided, the client will set a default list owned by the sender.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, res, unresolved};
use iota_types::{Address, Digest, ObjectId, ObjectReference, Transaction};

let client = iota_graphql_client::Client::new_devnet();
let sender =
    Address::from_str("0x611830d3641a68f94a690dcc25d1f4b0dac948325ac18f6dd32564371735f32c")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);

let gas_coin1 =
    ObjectId::from_str("0x0b0270ee9d27da0db09651e5f7338dfa32c7ee6441ccefa1f6e305735bcfc7ab")?;
let gas_coin2 =
    ObjectId::from_str("0xd04077fe3b6fad13b3d4ed0d535b7ca92afcac8f0f2a0e0925fb9f4f0b30c699")?;

builder
    .split_coins(unresolved::Argument::Gas, [1000u64])
    .gas([gas_coin1, gas_coin2]);

let txn: Transaction = builder.finish().await?;

pub async fn finish(self) -> Result<Transaction, Error>

Convert this builder into a transaction.

pub async fn dry_run(self, skip_checks: bool) -> Result<DryRunResult, Error>

Dry run the transaction.

pub async fn execute( self, keypair: &SimpleKeypair, wait_for: impl Into<Option<WaitForTx>>, ) -> Result<TransactionEffects, Error>

Execute the transaction and optionally wait for finalization. The GraphQL client will be used unless a gas station was configured, in which case the transaction will be sent to the endpoint for execution.

pub async fn execute_with_sponsor( self, keypair: &SimpleKeypair, sponsor_keypair: &SimpleKeypair, wait_for: impl Into<Option<WaitForTx>>, ) -> Result<TransactionEffects, Error>

Execute the transaction with a sponsor keypair and optionally wait for finalization.

impl<C> TransactionBuilder<C, MoveCall>

pub fn arguments<U: PTBArgumentList>(&mut self, params: U) -> &mut Self

Set the call params. Optional.

impl<C> TransactionBuilder<C, MoveCall>

pub fn generics<G: MoveTypes>(&mut self) -> &mut Self

Set the generic type arguments. Optional.

pub fn type_tags( &mut self, tags: impl IntoIterator<Item = TypeTag>, ) -> &mut Self

Set the type arguments manually. Optional.

impl TransactionBuilder<(), Publish>

pub fn package_id(&mut self, name: impl NamedResult) -> &mut TransactionBuilder

Get the package ID from the UpgradeCap so that it can be used for future commands.

impl<C: ClientMethods> TransactionBuilder<C, Publish>

pub fn package_id( &mut self, name: impl NamedResult, ) -> &mut TransactionBuilder<C>

Get the package ID from the UpgradeCap so that it can be used for future commands.

impl<C> TransactionBuilder<C, Publish>

pub fn upgrade_cap( &mut self, name: impl NamedResult, ) -> &mut TransactionBuilder<C>

Finish the publish call and return the UpgradeCap.

impl<C, L: Into<Command>> TransactionBuilder<C, L>

pub fn name(&mut self, name: impl NamedResults) -> &mut Self

Set the name for the last command.

pub fn arg(&mut self) -> Argument

Get the argument representing the last command.

impl<C> TransactionBuilder<C, GasStationData>

pub fn gas_reservation_duration(&mut self, duration: Duration) -> &mut Self

Set the gas reservation duration for a gas station sponsor.

pub fn add_gas_station_header( &mut self, name: HeaderName, value: HeaderValue, ) -> &mut Self

Add a header that will be passed to the gas station sponsor request.