iota_sdk_transaction_builder

Trait ClientMethods

pub trait ClientMethods {
    type Error: 'static + Error + Send + Sync;

    // Required methods
    fn object(
        &self,
        object_id: ObjectId,
        version: impl Into<Option<u64>>,
    ) -> impl Future<Output = Result<Option<Object>, Self::Error>>;
    fn objects(
        &self,
        type_tag: Option<TypeTag>,
        owner: Option<Address>,
        object_ids: Option<Vec<ObjectId>>,
        ascending: bool,
        cursor: Option<String>,
        limit: Option<usize>,
    ) -> impl Future<Output = Result<Vec<Object>, Self::Error>>;
    fn transaction(
        &self,
        digest: Digest,
    ) -> impl Future<Output = Result<Option<SignedTransaction>, Self::Error>>;
    fn transaction_effects(
        &self,
        digest: Digest,
    ) -> impl Future<Output = Result<Option<TransactionEffects>, Self::Error>>;
    fn reference_gas_price(
        &self,
        epoch: impl Into<Option<u64>>,
    ) -> impl Future<Output = Result<Option<u64>, Self::Error>>;
    fn estimate_tx_budget(
        &self,
        tx: &Transaction,
    ) -> impl Future<Output = Result<Option<u64>, Self::Error>>;
    fn dry_run_tx(
        &self,
        tx: &Transaction,
        skip_checks: bool,
    ) -> impl Future<Output = Result<DryRunResult, Self::Error>>;
    fn execute_tx(
        &self,
        signatures: &[UserSignature],
        tx: &Transaction,
        wait_for: impl Into<Option<WaitForTx>>,
    ) -> impl Future<Output = Result<TransactionEffects, Self::Error>>;
    fn wait_for_tx(
        &self,
        digest: Digest,
        wait_for: WaitForTx,
    ) -> impl Future<Output = Result<(), Self::Error>>;
}

A trait which defines methods needed from the client for the Transaction Builder.

Required Associated Types

type Error: 'static + Error + Send + Sync

The error type for this client.

Required Methods

fn object( &self, object_id: ObjectId, version: impl Into<Option<u64>>, ) -> impl Future<Output = Result<Option<Object>, Self::Error>>

Fetch an object

fn objects( &self, type_tag: Option<TypeTag>, owner: Option<Address>, object_ids: Option<Vec<ObjectId>>, ascending: bool, cursor: Option<String>, limit: Option<usize>, ) -> impl Future<Output = Result<Vec<Object>, Self::Error>>

Fetch objects

fn transaction( &self, digest: Digest, ) -> impl Future<Output = Result<Option<SignedTransaction>, Self::Error>>

Fetch a transaction

fn transaction_effects( &self, digest: Digest, ) -> impl Future<Output = Result<Option<TransactionEffects>, Self::Error>>

Fetch transaction effects

fn reference_gas_price( &self, epoch: impl Into<Option<u64>>, ) -> impl Future<Output = Result<Option<u64>, Self::Error>>

Get the reference gas price

fn estimate_tx_budget( &self, tx: &Transaction, ) -> impl Future<Output = Result<Option<u64>, Self::Error>>

Estimate the gas budget needed for a transaction

fn dry_run_tx( &self, tx: &Transaction, skip_checks: bool, ) -> impl Future<Output = Result<DryRunResult, Self::Error>>

Dry run a transaction

fn execute_tx( &self, signatures: &[UserSignature], tx: &Transaction, wait_for: impl Into<Option<WaitForTx>>, ) -> impl Future<Output = Result<TransactionEffects, Self::Error>>

Execute a transaction

fn wait_for_tx( &self, digest: Digest, wait_for: WaitForTx, ) -> impl Future<Output = Result<(), Self::Error>>

Wait for the indexing or finalization of a transaction by its digest.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.