iota_sdk_transaction_builder

Module types

Types for use with the transaction builder.

Structs

PureBytes

Pure BCS bytes

Traits

MoveArg

A trait which defines how types are serialized for move calls.

MoveArgCollection

A trait which defines how collections of move arg types are serialized for move calls.

MoveType

A trait which defines the tag of the type.

MoveTypes

A trait which defines multiple types for use with tuples.