iota_sdk_transaction_builder::builder

Struct TransactionBuilder

pub struct TransactionBuilder<C = (), L = ()> { /* private fields */ }

A transaction builder which can be used to construct [Transaction]s.

Implementations

impl TransactionBuilder

pub fn new(sender: Address) -> Self

Instantiate a new PTB.

pub fn with_client<C>(self, client: C) -> TransactionBuilder<C>

Set the client to enable automatic object resolution.

impl<C, L> TransactionBuilder<C, L>

pub fn sender(&mut self, sender: Address) -> &mut Self

Set the sender address.

pub fn apply_argument<P: PTBArgument>(&mut self, param: P) -> Argument

Apply the given parameter and return the generated argument

pub fn apply_arguments<P: PTBArgumentList>(&mut self, param: P) -> Vec<Argument>

Apply the given parameters and return the generated arguments

pub fn get_gas(&self) -> Vec<ObjectId>

Get the current set gas coins.

pub fn gas_budget(&mut self, gas_budget: u64) -> &mut Self

Set the gas budget. Optional.

pub fn gas_price(&mut self, gas_price: u64) -> &mut Self

Set the gas price. Optional.

pub fn sponsor(&mut self, sponsor: Address) -> &mut Self

Set the sponsor. Optional.

pub fn gas_station_sponsor( &mut self, url: Url, ) -> &mut TransactionBuilder<C, GasStationData>

Set the gas station sponsor. Optional.

pub fn expiration(&mut self, expiration: u64) -> &mut Self

Set the expiration. Optional.

pub fn input(&mut self, input: Input) -> Argument

Make a value available to the transaction as an input.

pub fn pure_bytes(&mut self, bytes: Vec<u8>) -> Argument

Add a pure input using the BCS serialized bytes

pub fn pure<T: Serialize>(&mut self, value: T) -> Argument

Add a pure input

pub fn command(&mut self, command: Command) -> Argument

Add a new command to the PTB

pub fn assigned_command(&mut self, cmd: Command, name: impl AssignedResults)

Manually set a command with an optional name

pub fn get_assigned_result(&self, name: &str) -> Option<Argument>

Get the value for the given string in the assigned results map

pub fn move_call( &mut self, package_id: impl Into<ObjectId>, module: &str, function: &str, ) -> &mut TransactionBuilder<C, MoveCall>

Begin building a move call.

pub fn transfer_objects<U: PTBArgumentList>( &mut self, recipient: Address, objects: U, ) -> &mut TransactionBuilder<C>

Transfer objects to a recipient address.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, assigned};
use iota_types::{Address, ObjectDigest, ObjectId, ObjectReference, Transaction, Version};

let sender =
    Address::from_str("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);

builder.transfer_objects(
    Address::from_str("0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900")?,
    (
        // ObjectIds can be passed when a client is provided
        ObjectId::from_str(
            "0x65beb18e282d1f33a39bffa84ff92ec4d2fec0350ba6f7e5a568afff72d651db",
        )?,
        // ObjectReferences are always allowed, though they must be correct
        ObjectReference {
            object_id: ObjectId::from_str(
                "0xe0e45ecb12ddca5f0d5192d2ee9e7f711959aa98614f9905e1e25c612ffd99a2",
            )?,
            digest: ObjectDigest::from_str("hSAGU3ZwDwxptd17ZK1QPDdJLhvPMfpSxe1p892GFVn")?,
            version: Version::from_u64(545110774),
        },
        // The result of a previous command can also be used
        assigned("coin"),
    ),
);

let txn: Transaction = builder.finish().await?;

pub fn send_iota<T: PTBArgument>( &mut self, recipient: Address, amount: T, ) -> &mut TransactionBuilder<C>

Send IOTA to a recipient address.

The amount parameter specifies the quantity in NANOS, where 1 IOTA equals 1_000_000_000 NANOS. That amount is split from the gas coin and sent.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::Address;

let from_address =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let to_address =
    Address::from_hex("0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900")?;

let mut builder = TransactionBuilder::new(from_address).with_client(client);
builder.send_iota(to_address, 5000000000u64);
let txn = builder.finish().await?;

pub fn send_coins<T: PTBArgumentList, U: PTBArgument>( &mut self, coins: T, recipient: Address, amount: impl Into<Option<U>>, ) -> &mut TransactionBuilder<C>

Transfer some coins to a recipient address. If multiple coins are provided then they will be merged.

The amount parameter specifies the quantity in NANOS, where 1 IOTA equals 1_000_000_000 NANOS. If amount is provided, that amount is split from the provided coins and sent. If amount is None, the entire coins are transferred.

All provided coins must have the same coin type. Mixing coins of different types will result in an error.

If you intend to transfer all provided coins to another address in a single transaction, consider using TransactionBuilder::transfer_objects() instead.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let from_address =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let to_address =
    Address::from_hex("0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900")?;

// This is a coin of type
// 0xfce9c14e5f0c2b65787debb8145a33a4a2fc83152e8939000b862e174bc86bb8::cert::CERT
let coin =
    ObjectId::from_hex("0xe0e45ecb12ddca5f0d5192d2ee9e7f711959aa98614f9905e1e25c612ffd99a2")?;

let mut builder = TransactionBuilder::new(from_address).with_client(client);
builder.send_coins([coin], to_address, 50000000000u64);
let txn = builder.finish().await?;

pub fn pay<T: PTBArgumentList, U: PTBArgument>( &mut self, coins: T, payments: impl IntoIterator<Item = (Address, U)>, ) -> &mut TransactionBuilder<C>

Send coins to multiple recipients, each paired with the amount to send.

The amounts specify quantities in the coins’ smallest unit (NANOS for IOTA coins, where 1 IOTA equals 1_000_000_000 NANOS).

The coins are merged into the first one, the amounts are split off it in a single command, and each split coin is transferred to its corresponding recipient, with one transfer command per unique recipient. The remainder stays in the first coin.

All provided coins must have the same coin type. Mixing coins of different types will result in an error.

Adds no commands if either list is empty.

To pay IOTA directly from the gas coin, use TransactionBuilder::pay_iota(), or pass unresolved::Argument::Gas as the only coin.

For a single recipient, consider using TransactionBuilder::send_coins() or TransactionBuilder::send_iota() instead.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let from_address =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let payments = [
    (
        Address::from_hex(
            "0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900",
        )?,
        50000000000u64,
    ),
    (
        Address::from_hex(
            "0x111ff11c96e2c9b19d2c47ab973012483b136dfbfee55f79b32ed4980e6ef11c",
        )?,
        25000000000,
    ),
];

// This is a coin of type
// 0xfce9c14e5f0c2b65787debb8145a33a4a2fc83152e8939000b862e174bc86bb8::cert::CERT
let coin =
    ObjectId::from_hex("0xe0e45ecb12ddca5f0d5192d2ee9e7f711959aa98614f9905e1e25c612ffd99a2")?;

let mut builder = TransactionBuilder::new(from_address).with_client(client);
builder.pay([coin], payments);
let txn = builder.finish().await?;

pub fn pay_iota<U: PTBArgument>( &mut self, payments: impl IntoIterator<Item = (Address, U)>, ) -> &mut TransactionBuilder<C>

Send IOTA to multiple recipients, each paired with the amount to send.

The amounts specify quantities in NANOS, where 1 IOTA equals 1_000_000_000 NANOS. They are split off the gas coin in a single command, and each split coin is transferred to its corresponding recipient, with one transfer command per unique recipient.

To pay with specific coins, or with a coin type other than IOTA, use TransactionBuilder::pay(). For a single recipient, consider using TransactionBuilder::send_iota() instead.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::Address;

let from_address =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let payments = [
    (
        Address::from_hex(
            "0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900",
        )?,
        50000000000u64,
    ),
    (
        Address::from_hex(
            "0x111ff11c96e2c9b19d2c47ab973012483b136dfbfee55f79b32ed4980e6ef11c",
        )?,
        25000000000,
    ),
];

let mut builder = TransactionBuilder::new(from_address).with_client(client);
builder.pay_iota(payments);
let txn = builder.finish().await?;

pub fn merge_coins<T: PTBArgument, U: PTBArgumentList>( &mut self, primary_coin: T, consumed_coins: U, ) -> &mut TransactionBuilder<C>

Merge multiple coins into one.

This method combines the balances of multiple coins of the same coin type into a single coin. The primary_coin will receive the balances from all consumed_coins. After merging, the consumed_coins will be consumed and no longer exist.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let sender =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;

let coin_0 =
    ObjectId::from_hex("0xdc956de89b914e6a7fbd83caebefc8ec91be1207667ea5576386391aa82449cc")?;
let coin_1 =
    ObjectId::from_hex("0x65beb18e282d1f33a39bffa84ff92ec4d2fec0350ba6f7e5a568afff72d651db")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.merge_coins(coin_0, [coin_1]);
let txn = builder.finish().await?;

pub fn split_coins<T: PTBArgument, U: PTBArgumentList>( &mut self, coin: T, split_amounts: U, ) -> &mut TransactionBuilder<C, SplitCoins>

Split a coin into many.

Example

use iota_sdk_transaction_builder::{TransactionBuilder, assigned};
use iota_types::{Address, ObjectId};

let sender =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let coin =
    ObjectId::from_hex("0xdc956de89b914e6a7fbd83caebefc8ec91be1207667ea5576386391aa82449cc")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder
    .split_coins(coin, [1000u64, 2000, 3000])
    .assign(("coin1", "coin2", "coin3"))
    .transfer_objects(
        sender,
        (assigned("coin1"), assigned("coin2"), assigned("coin3")),
    );
let txn = builder.finish().await?;

pub fn publish_package( &mut self, package_data: MovePackageData, ) -> &mut TransactionBuilder<C, Publish>

Publish a move package.

pub fn authorize_upgrade<Cap: PTBArgument, P: PTBArgument, D: PTBArgument>( &mut self, upgrade_capability: Cap, upgrade_policy: P, digest: D, ) -> &mut TransactionBuilder<C, MoveCall>

Authorize a package upgrade by calling 0x2::package::authorize_upgrade.

The command’s result is the UpgradeTicket expected by upgrade. The policy is one of the UpgradePolicy constants and the digest is the digest of the package being upgraded to ([MovePackageData::digest]).

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, MovePackageData, ObjectId, UpgradePolicy};

let sender =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let package_id =
    ObjectId::from_hex("0xdc956de89b914e6a7fbd83caebefc8ec91be1207667ea5576386391aa82449cc")?;
let upgrade_cap =
    ObjectId::from_hex("0x13e8d0b5cdec156e44c0834274a431505954eb27a8f774a7f9044c2908c1c494")?;
let package_data = MovePackageData::new(modules, dependencies);

let mut builder = TransactionBuilder::new(sender).with_client(client);
let upgrade_ticket = builder
    .authorize_upgrade(upgrade_cap, UpgradePolicy::COMPATIBLE, package_data.digest)
    .result();
let upgrade_receipt = builder
    .upgrade(package_id, package_data, upgrade_ticket)
    .result();
builder.commit_upgrade(upgrade_cap, upgrade_receipt);
let txn = builder.finish().await?;

pub fn upgrade<U: PTBArgument>( &mut self, package_id: ObjectId, package_data: MovePackageData, upgrade_ticket: U, ) -> &mut TransactionBuilder<C, Upgrade>

Add the raw Upgrade command, consuming the UpgradeTicket returned by authorize_upgrade and producing the UpgradeReceipt expected by commit_upgrade.

For the standard upgrade flow, use upgrade_package instead.

pub fn commit_upgrade<Cap: PTBArgument, R: PTBArgument>( &mut self, upgrade_capability: Cap, upgrade_receipt: R, ) -> &mut TransactionBuilder<C, MoveCall>

Commit a package upgrade by calling 0x2::package::commit_upgrade, consuming the UpgradeReceipt returned by upgrade.

See authorize_upgrade for an example of the full upgrade flow.

pub fn upgrade_package<Cap: PTBArgument, P: PTBArgument>( &mut self, package_id: ObjectId, package_data: MovePackageData, upgrade_capability: Cap, upgrade_policy: P, ) -> &mut TransactionBuilder<C, MoveCall>

Upgrade a move package.

This is a high-level function which chains authorize_upgrade, upgrade and commit_upgrade using the digest from the provided [MovePackageData]. Use the individual functions for custom upgrade flows.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, MovePackageData, ObjectId, UpgradePolicy};

let sender =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let package_id =
    ObjectId::from_hex("0xdc956de89b914e6a7fbd83caebefc8ec91be1207667ea5576386391aa82449cc")?;
let upgrade_cap =
    ObjectId::from_hex("0x13e8d0b5cdec156e44c0834274a431505954eb27a8f774a7f9044c2908c1c494")?;
let package_data = MovePackageData::new(modules, dependencies);

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.upgrade_package(
    package_id,
    package_data,
    upgrade_cap,
    UpgradePolicy::COMPATIBLE,
);
let txn = builder.finish().await?;

pub fn stake<S: PTBArgument>( &mut self, stake_amount: S, validator_address: Address, ) -> &mut Self

Add stake to a validator’s staking pool.

This is a high-level function which will split the provided stake amount from the gas coin and then stake using the resulting coin.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let sender =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.stake(1000000000u64, validator_address);
let txn = builder.finish().await?;

pub fn unstake<S: PTBArgument>(&mut self, staked_iota: S) -> &mut Self

Withdraw stake from a validator’s staking pool.

Example

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId};

let sender =
    Address::from_hex("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
// This is a 0x3::staking_pool::StakedIota owned by the sender
let staked_iota =
    ObjectId::from_hex("0x13e8d0b5cdec156e44c0834274a431505954eb27a8f774a7f9044c2908c1c494")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.unstake(staked_iota);
let txn = builder.finish().await?;

pub fn make_move_vec<T: PTBArgument + MoveType>( &mut self, elements: impl IntoIterator<Item = T>, ) -> &mut TransactionBuilder<C, MakeMoveVector>

Make a move vector from a list of elements.

Often it is possible (and more efficient) to pass a rust slice or Vec instead of calling this function, which will serialize the bytes into a move vector pure argument.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, assigned};
use iota_types::{Address, Transaction};

let sender = "0x71b4b4f171b4355ff691b7c470579cf1a926f96f724e5f9a30efc4b5f75d085e".parse()?;

let mut builder = TransactionBuilder::new(sender).with_client(client);

let address1 =
    Address::from_str("0xde49ea53fbadee67d3e35a097cdbea210b659676fc680a0b0c5f11d0763d375e")?;
let address2 =
    Address::from_str("0xe512234aa4ef6184c52663f09612b68f040dd0c45de037d96190a071ca5525b3")?;

builder
    .make_move_vec([address1, address2])
    .assign("addresses")
    .move_call(Address::FRAMEWORK, "vec_map", "from_keys_values")
    .generics::<(Address, u64)>()
    .arguments((assigned("addresses"), [10000000u64, 20000000u64]));

let txn: Transaction = builder.finish().await?;

impl<L> TransactionBuilder<(), L>

pub fn gas( &mut self, obj_refs: impl IntoIterator<Item = ObjectReference>, ) -> &mut Self

Add gas coins that will be consumed. Optional.

A gas coin is paid as gas instead of being passed as an input, so it cannot also be passed to a command — building the transaction fails if one is. The exception is transferring every gas coin to the same recipient in a single transfer_objects: the coins are smashed into one before the commands run, and that coin is transferred with whatever balance is left after gas.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, assigned, unresolved};
use iota_types::{Address, ObjectDigest, ObjectId, ObjectReference, Transaction, Version};

let sender =
    Address::from_str("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;

let mut builder = TransactionBuilder::new(sender);

let gas_coin1 = ObjectReference {
    object_id: ObjectId::from_str(
        "0xdc956de89b914e6a7fbd83caebefc8ec91be1207667ea5576386391aa82449cc",
    )?,
    digest: ObjectDigest::from_str("CPpQZqyHZcG2Pb9gZyikbc8dEuyipXHR6ihnfe9iYiMt")?,
    version: Version::from_u64(473053811),
};
let gas_coin2 = ObjectReference {
    object_id: ObjectId::from_str(
        "0x65beb18e282d1f33a39bffa84ff92ec4d2fec0350ba6f7e5a568afff72d651db",
    )?,
    digest: ObjectDigest::from_str("8ahH5RXFnK1jttQEWTypYX7MRzLuQDEXk7fhMHCyZekX")?,
    version: Version::from_u64(473053810),
};

builder
    .split_coins(unresolved::Argument::Gas, [1000u64])
    .gas([gas_coin1, gas_coin2])
    .gas_budget(1000000000)
    .gas_price(100);

let txn: Transaction = builder.finish()?;

pub fn finish(self) -> Result<Transaction, Error>

Convert this builder into a transaction.

pub async fn execute_with_gas_station( self, signer: &impl TransactionSigner, ) -> Result<Value, Error>

Execute the transaction using the gas station and return the JSON transaction effects. This will fail unless data is set with Self::gas_station_sponsor.

NOTE: These effects are not necessarily compatible with [TransactionEffects]

impl<C, L> TransactionBuilder<C, L>

pub fn get_client(&self) -> &C

Get the client used by the builder.

impl<C: TransactionBuilderLedgerClient, L> TransactionBuilder<C, L>

pub fn gas(&mut self, obj_ids: impl IntoIterator<Item = ObjectId>) -> &mut Self

Add gas coins that will be consumed. If no gas coins are provided, the client will set a default list owned by the sender.

A gas coin is paid as gas instead of being passed as an input, so it cannot also be passed to a command — building the transaction fails if one is. The exception is transferring every gas coin to the same recipient in a single transfer_objects: the coins are smashed into one before the commands run, and that coin is transferred with whatever balance is left after gas.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, assigned, unresolved};
use iota_types::{Address, ObjectDigest, ObjectId, ObjectReference, Transaction};

let sender =
    Address::from_str("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);

let gas_coin1 =
    ObjectId::from_str("0xdc956de89b914e6a7fbd83caebefc8ec91be1207667ea5576386391aa82449cc")?;
let gas_coin2 =
    ObjectId::from_str("0x65beb18e282d1f33a39bffa84ff92ec4d2fec0350ba6f7e5a568afff72d651db")?;

builder
    .split_coins(unresolved::Argument::Gas, [1000u64])
    .gas([gas_coin1, gas_coin2]);

let txn: Transaction = builder.finish().await?;

pub fn gas_refs( &mut self, obj_refs: impl IntoIterator<Item = ObjectReference>, ) -> &mut Self

Add gas coins that will be consumed, by reference. Optional.

Same as gas, but for coins the caller has already resolved: the builder takes the references as given instead of looking the objects up. Setting either one stops the builder from picking gas coins on its own.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::{TransactionBuilder, unresolved};
use iota_types::{Address, ObjectDigest, ObjectId, ObjectReference, Transaction, Version};

let sender =
    Address::from_str("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);

let gas_coin1 = ObjectReference {
    object_id: ObjectId::from_str(
        "0xdc956de89b914e6a7fbd83caebefc8ec91be1207667ea5576386391aa82449cc",
    )?,
    digest: ObjectDigest::from_str("CPpQZqyHZcG2Pb9gZyikbc8dEuyipXHR6ihnfe9iYiMt")?,
    version: Version::from_u64(473053811),
};
let gas_coin2 = ObjectReference {
    object_id: ObjectId::from_str(
        "0x65beb18e282d1f33a39bffa84ff92ec4d2fec0350ba6f7e5a568afff72d651db",
    )?,
    digest: ObjectDigest::from_str("8ahH5RXFnK1jttQEWTypYX7MRzLuQDEXk7fhMHCyZekX")?,
    version: Version::from_u64(473053810),
};

builder
    .split_coins(unresolved::Argument::Gas, [1000u64])
    .gas_refs([gas_coin1, gas_coin2])
    .gas_budget(1000000000)
    .gas_price(100);

let txn: Transaction = builder.finish().await?;

pub async fn finish_with_budget( self, gas_budget: u64, ) -> Result<Transaction, Error>

Convert this builder into a transaction with the given gas budget, used as-is (no estimation or minimum clamp) and overriding any budget set via gas_budget.

pub async fn finish_kind(self) -> Result<TransactionKind, Error>

Convert this builder into a [TransactionKind], resolving the inputs with the client but leaving the gas alone.

Use this when the gas payment is decided elsewhere — a wallet that picks the gas coins itself, or a caller that needs the kind to estimate a budget before it can pick them. Unlike finish, no gas coins are selected, no budget is estimated and no gas price is fetched, so this costs nothing beyond resolving the inputs. Any gas set with gas or gas_refs, and the sponsor, price, budget and expiration, are not part of a [TransactionKind] and are dropped.

Example

use std::str::FromStr;

use iota_sdk_transaction_builder::TransactionBuilder;
use iota_types::{Address, ObjectId, TransactionKind};

let sender =
    Address::from_str("0xda1820edf693ee32b5729907b9b2ec8e64980ee8c008c17e89cfb4e5ecd72151")?;
let to_address =
    Address::from_str("0x0000a4984bd495d4346fa208ddff4f5d5e5ad48c21dec631ddebc99809f16900")?;
let coin =
    ObjectId::from_str("0xe0e45ecb12ddca5f0d5192d2ee9e7f711959aa98614f9905e1e25c612ffd99a2")?;

let mut builder = TransactionBuilder::new(sender).with_client(client);
builder.transfer_objects(to_address, [coin]);

let kind: TransactionKind = builder.finish_kind().await?;

impl<C: TransactionBuilderLedgerClient + TransactionBuilderSimulationClient, L> TransactionBuilder<C, L>

pub async fn finish(self) -> Result<Transaction, Error>

Convert this builder into a transaction.

When no gas budget was set, it is estimated by simulating the transaction via the client (estimate_transaction_budget). To build with an explicit budget on a client without simulation support, use finish_with_budget.

pub async fn dry_run(self, skip_checks: bool) -> Result<C::DryRunResult, Error>

Dry run the transaction.

impl<C: TransactionBuilderClient, L> TransactionBuilder<C, L>

pub async fn execute( self, signer: &impl TransactionSigner, wait_for: impl Into<Option<WaitForTransaction>>, ) -> Result<TransactionEffects, Error>

Execute the transaction and optionally wait for finalization. The client will be used unless a gas station was configured, in which case the transaction will be sent to the endpoint for execution.

pub async fn execute_with_sponsor( self, signer: &impl TransactionSigner, sponsor_signer: &impl TransactionSigner, wait_for: impl Into<Option<WaitForTransaction>>, ) -> Result<TransactionEffects, Error>

Execute the transaction with a sponsor signer and optionally wait for finalization.

impl<C> TransactionBuilder<C, MoveCall>

pub fn arguments<U: PTBArgumentList>(&mut self, params: U) -> &mut Self

Set the call params. Optional.

impl<C> TransactionBuilder<C, MoveCall>

pub fn generics<G: MoveTypes>(&mut self) -> &mut Self

Set the generic type arguments. Optional.

pub fn type_tags( &mut self, tags: impl IntoIterator<Item = TypeTag>, ) -> &mut Self

Set the type arguments manually. Optional.

impl<C> TransactionBuilder<C, Publish>

pub fn package_id( &mut self, name: impl AssignedResult, ) -> &mut TransactionBuilder<C>

Get the package ID from the UpgradeCap so that it can be used for future commands.

pub fn upgrade_cap( &mut self, name: impl AssignedResult, ) -> &mut TransactionBuilder<C>

Finish the publish call and return the UpgradeCap.

impl<C, L: Into<Command>> TransactionBuilder<C, L>

pub fn assign(&mut self, name: impl AssignedResults) -> &mut Self

Assign a name to the last command’s result.

pub fn result(&mut self) -> Argument

Get the argument representing the last command.

impl<C> TransactionBuilder<C, GasStationData>

pub fn gas_reservation_duration(&mut self, duration: Duration) -> &mut Self

Set the gas reservation duration for a gas station sponsor.

pub fn add_gas_station_header( &mut self, name: HeaderName, value: HeaderValue, ) -> &mut Self

Add a header that will be passed to the gas station sponsor request.